# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fogos.pt",
        "contact": {
            "name": "Tomahock",
            "email": "mail@fogos.pt"
        },
        "version": "1.0"
    },
    "paths": {
        "/new/fires": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "description": "Active Fires",
                "operationId": "App\\Http\\Controllers\\LegacyController::newFires",
                "responses": {
                    "default": {
                        "description": "Active Fires"
                    }
                }
            }
        },
        "/fires": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Fire detail",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::fires",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "FireId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "fire id"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/warnings": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Warnings list",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::warnings",
                "responses": {
                    "default": {
                        "description": "Warnings list"
                    }
                }
            }
        },
        "/v1/madeira/warnings": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Madeira Warnings list",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::warningsMadeira",
                "responses": {
                    "default": {
                        "description": "Madeira Warnings list"
                    }
                }
            }
        },
        "/v1/stats/week": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Last 7 days total fires",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::statsWeek",
                "responses": {
                    "default": {
                        "description": "Last 7 days total fires"
                    }
                }
            }
        },
        "/fires/danger": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Fire RCM",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::firesDanger",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "FireId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Fire RCM"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/fires/status": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Fire Status History",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::firesStatus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "FireId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Fire Status History"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/madeira/fires": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "NOT WORKING",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::firesMadeira",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "FireId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "NOT WORKING"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/madeira/fires/status": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "NOT WORKING",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::firesStatusMadeira",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "FireId",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "NOT WORKING"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/v1/now": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Active fires status now",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::now",
                "responses": {
                    "default": {
                        "description": "Active fires status now"
                    }
                }
            }
        },
        "/v1/stats/8hours": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Today district stats for 00H to 08H and 08H to 16H",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::stats8hours",
                "responses": {
                    "default": {
                        "description": "Today district stats for 00H to 08H and 08H to 16H"
                    }
                }
            }
        },
        "/v1/stats/8hours/yesterday": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Yesterday district stats for 00H to 08H, 08H to 16H and 16H to 24H",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::stats8hoursYesterday",
                "responses": {
                    "default": {
                        "description": "Yesterday district stats for 00H to 08H, 08H to 16H and 16H to 24H"
                    }
                }
            }
        },
        "/v1/stats/last-night": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Last night fires (21H to 09H today)",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::lastNight",
                "responses": {
                    "default": {
                        "description": "Last night fires (21H to 09H today)"
                    }
                }
            }
        },
        "/v1/status": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Current status in text",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::status",
                "responses": {
                    "default": {
                        "description": "Current status in text"
                    }
                }
            }
        },
        "/v1/active": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Active fire list",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::active",
                "responses": {
                    "default": {
                        "description": "Active fire list"
                    }
                }
            }
        },
        "/v1/aerial": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Meios aereos activos em texto",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::aerial",
                "responses": {
                    "default": {
                        "description": "Meios aereos activos em texto"
                    }
                }
            }
        },
        "/v1/stats": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Stats district today text",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::stats",
                "responses": {
                    "default": {
                        "description": "Stats district today text"
                    }
                }
            }
        },
        "/v1/risk": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "RCM for concelho",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::risk",
                "parameters": [
                    {
                        "name": "concelho",
                        "in": "query",
                        "description": "Concelho",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "RCM for concelho"
                    }
                }
            }
        },
        "/v1/list": {
            "get": {
                "tags": [
                    "legacy"
                ],
                "summary": "Active fires for concelho",
                "description": "",
                "operationId": "App\\Http\\Controllers\\LegacyController::listConcelho",
                "parameters": [
                    {
                        "name": "concelho",
                        "in": "query",
                        "description": "Concelho",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Active fires for concelho"
                    }
                }
            }
        }
    }
}